Fixing race condition on toFile
Closes #6 

`cb()` is already called even before the `toFile(...)` operation is complete, because `on('finish', ...)` runs regardless of how `toFile` is being executed. This causes file r/w race when the user loads the image file as soon as the controller returns the filename. Precisely speaking, the image file is being read while the image file is being written to the upload folder, making the image returning an empty or a truncated output.